add RLE block copy function

+++ src/dct.c
@@ -33,7 +33,7 @@ void dct_quantize_floatblock(
 		for (int32_t x=0;x<size;x++) {
 			float in = (float)input->data[y][x];
 
-			int32_t out = round(in/quant_matrix.data[y][x]);			
+			int32_t out = round(in/(float)quant_matrix.data[y][x]);			
 			output->data[y][x] = out;
 		}
 	}
